Compatible with new hardware layouts! -melgeek/z70ultra!
