+++ keyboards/melgeek/z70ultra/info.json
@@ -70,7 +70,9 @@
                 {"x":0, "y":4, "w":1.25},
                 {"x":1.25, "y":4, "w":1.25},
                 {"x":2.5, "y":4, "w":1.25},
+                {"x":3.75, "y":4, "w":0},
                 {"x":3.75, "y":4, "w":6.25},
+                {"x":10, "y":4, "w":0},
                 {"x":10, "y":4, "w":1.25},
                 {"x":11.25, "y":4, "w":1.25},
                 {"x":13, "y":4},
@@ -146,6 +148,7 @@
                 {"x":1.25, "y":4, "w":1.25},
                 {"x":2.5, "y":4, "w":1.25},
                 {"x":3.75, "y":4, "w":3},
+                {"x":6.75, "y":4, "w":0},
                 {"x":6.75, "y":4, "w":3},
                 {"x":9.75, "y":4, "w":1.5},
                 {"x":11.25, "y":4, "w":1.25},
