+++ keyboards/melgeek/z70ultra/rev2/config.h
@@ -18,14 +18,14 @@
 
 /* USB Device descriptor parameter */
 #define VENDOR_ID       0xEDED
-#define PRODUCT_ID      0x6570
+#define PRODUCT_ID      0x6575
 #define DEVICE_VER      0x0001
 #define MANUFACTURER    MelGeek
 #define PRODUCT         Z70Ultra
 
 /* key matrix size */
-#define MATRIX_ROWS 6
-#define MATRIX_COLS 14
+#define MATRIX_ROWS 5
+#define MATRIX_COLS 15
 
 /*
  * Keyboard Matrix Assignments
@@ -38,7 +38,7 @@
  *
  */
 
-#define MATRIX_ROW_PINS { B12, B13, B14, B15, A3, B9 }
-#define MATRIX_COL_PINS { B11, B10, B1, A10, B5, B4, B3, A15, A2, A1, A0, C15, C14, C13 }
+#define MATRIX_ROW_PINS { B12, B13, B14, B15, A3 }
+#define MATRIX_COL_PINS { B11, B10, B1, A10, B5, B4, B3, A15, A2, A1, A0, C15, C14, C13, B9 }
 #define UNUSED_PINS
