--- keyboards/melgeek/z70ultra/rules.mk
@@ -1,1 +1,0 @@
-DEFAULT_FOLDER = melgeek/z70ultra/rev1
